Refactor DcSimApp wiring (#12)
