+++ project/validationTests/3S1T/application.conf
@@ -28,7 +28,7 @@ dcsim {
   exportInputs = "validationTests/3S1T"
   exportTrainId = "T1"
   exportRunExcel = "project/validationTests/3S1T/T1/A-B.xlsx"
-  export.enabled = true
+  export.enabled = false
 
   # ---- Simulation control ----
   simulationControl {
