+++ project/validationTests/3S1T/dc/exports/nodes.csv
@@ -1,0 +1,6 @@
+name,section,track,position_m
+N0,1,1,100
+N1,1,1,0
+N2,1,1,1500
+N3,1,1,3000
+N99,1,1,3100
