+++ project/validationTests/3S1T/dc/exports/run.csv
@@ -1,0 +1,106 @@
+time_s,train_id,section,track,position_m,P_req_W
+0.0,T1,425.0,1.0,100.0,0.0
+1.0,T1,425.0,1.0,100.5,100000.0
+2.0,T1,425.0,1.0,102.00000000000001,200000.0
+3.0,T1,425.0,1.0,104.50000000000001,300000.0
+4.0,T1,425.0,1.0,108.00000000000001,400000.0
+5.0,T1,425.0,1.0,112.50000000000001,500000.0
+6.0,T1,425.0,1.0,118.00000000000003,600000.0
+7.0,T1,425.0,1.0,124.50000000000003,700000.0
+8.0,T1,425.0,1.0,132.00000000000003,800000.0
+9.0,T1,425.0,1.0,140.50000000000003,900000.0
+10.0,T1,425.0,1.0,150.00000000000006,1000000.0
+11.0,T1,425.0,1.0,160.50000000000006,1100000.0
+12.0,T1,425.0,1.0,172.00000000000006,1200000.0
+13.0,T1,425.0,1.0,184.50000000000009,1300000.0
+14.0,T1,425.0,1.0,198.00000000000009,1400000.0
+15.0,T1,425.0,1.0,212.5000000000001,1500000.0
+16.0,T1,425.0,1.0,228.00000000000009,1600000.0
+17.0,T1,425.0,1.0,244.5000000000001,1700000.0
+18.0,T1,425.0,1.0,262.0000000000001,1800000.0
+19.0,T1,425.0,1.0,280.5000000000001,1900000.0
+20.0,T1,425.0,1.0,300.00000000000017,2000000.0
+21.0,T1,425.0,1.0,320.50000000000017,2100000.0
+22.0,T1,425.0,1.0,342.00000000000017,2200000.0
+23.0,T1,425.0,1.0,364.5000000000002,2300000.0
+24.0,T1,425.0,1.0,388.0000000000002,2400000.0
+25.0,T1,425.0,1.0,412.5000000000002,0.0
+26.0,T1,425.0,1.0,437.5000000000003,0.0
+27.0,T1,425.0,1.0,462.5000000000003,0.0
+28.0,T1,425.0,1.0,487.50000000000034,0.0
+29.0,T1,425.0,1.0,512.5000000000003,0.0
+30.0,T1,425.0,1.0,537.5000000000003,0.0
+31.0,T1,425.0,1.0,562.5000000000003,0.0
+32.0,T1,425.0,1.0,587.5000000000003,0.0
+33.0,T1,425.0,1.0,612.5000000000003,0.0
+34.0,T1,425.0,1.0,637.5000000000005,0.0
+35.0,T1,425.0,1.0,662.5000000000005,0.0
+36.0,T1,425.0,1.0,687.5000000000005,0.0
+37.0,T1,425.0,1.0,712.5000000000005,0.0
+38.0,T1,425.0,1.0,737.5000000000005,0.0
+39.0,T1,425.0,1.0,762.5000000000005,0.0
+40.0,T1,425.0,1.0,787.5000000000006,0.0
+41.0,T1,425.0,1.0,812.5000000000006,0.0
+42.0,T1,425.0,1.0,837.5000000000006,0.0
+43.0,T1,425.0,1.0,862.5000000000006,0.0
+44.0,T1,425.0,1.0,887.5000000000006,0.0
+45.0,T1,425.0,1.0,912.5000000000007,0.0
+46.0,T1,425.0,1.0,937.5000000000007,0.0
+47.0,T1,425.0,1.0,962.5000000000007,0.0
+48.0,T1,425.0,1.0,987.5000000000007,0.0
+49.0,T1,425.0,1.0,1012.5000000000006,0.0
+50.0,T1,425.0,1.0,1037.5000000000005,0.0
+51.0,T1,425.0,1.0,1062.5000000000005,0.0
+52.0,T1,425.0,1.0,1087.5000000000005,0.0
+53.0,T1,425.0,1.0,1112.5000000000002,0.0
+54.0,T1,425.0,1.0,1137.5000000000002,0.0
+55.0,T1,425.0,1.0,1162.5,0.0
+56.0,T1,425.0,1.0,1187.5,0.0
+57.0,T1,425.0,1.0,1212.5,0.0
+58.0,T1,425.0,1.0,1237.4999999999998,0.0
+59.0,T1,425.0,1.0,1262.4999999999998,0.0
+60.0,T1,425.0,1.0,1287.4999999999995,0.0
+61.0,T1,425.0,1.0,1312.4999999999995,0.0
+62.0,T1,425.0,1.0,1337.4999999999995,0.0
+63.0,T1,425.0,1.0,1362.4999999999993,0.0
+64.0,T1,425.0,1.0,1387.4999999999993,0.0
+65.0,T1,425.0,1.0,1412.499999999999,0.0
+66.0,T1,425.0,1.0,1437.499999999999,0.0
+67.0,T1,425.0,1.0,1462.499999999999,0.0
+68.0,T1,425.0,1.0,1487.4999999999989,0.0
+69.0,T1,425.0,1.0,1512.4999999999989,0.0
+70.0,T1,425.0,1.0,1537.4999999999989,0.0
+71.0,T1,425.0,1.0,1562.4999999999986,0.0
+72.0,T1,425.0,1.0,1587.4999999999986,0.0
+73.0,T1,425.0,1.0,1612.4999999999984,0.0
+74.0,T1,425.0,1.0,1637.4999999999984,0.0
+75.0,T1,425.0,1.0,1662.4999999999984,0.0
+76.0,T1,425.0,1.0,1687.4999999999982,0.0
+77.0,T1,425.0,1.0,1712.4999999999982,0.0
+78.0,T1,425.0,1.0,1737.499999999998,0.0
+79.0,T1,425.0,1.0,1762.499999999998,2500000.0
+80.0,T1,425.0,1.0,1786.999999999998,2400000.0
+81.0,T1,425.0,1.0,1810.499999999998,2300000.0
+82.0,T1,425.0,1.0,1832.999999999998,2200000.0
+83.0,T1,425.0,1.0,1854.499999999998,2100000.0
+84.0,T1,425.0,1.0,1874.999999999998,2000000.0
+85.0,T1,425.0,1.0,1894.4999999999982,1900000.0
+86.0,T1,425.0,1.0,1912.999999999998,1800000.0
+87.0,T1,425.0,1.0,1930.4999999999982,1700000.0
+88.0,T1,425.0,1.0,1946.999999999998,1600000.0
+89.0,T1,425.0,1.0,1962.4999999999982,1500000.0
+90.0,T1,425.0,1.0,1976.9999999999982,1400000.0
+91.0,T1,425.0,1.0,1990.4999999999982,1300000.0
+92.0,T1,425.0,1.0,2002.9999999999984,1200000.0
+93.0,T1,425.0,1.0,2014.4999999999982,1100000.0
+94.0,T1,425.0,1.0,2024.9999999999982,1000000.0
+95.0,T1,425.0,1.0,2034.4999999999982,900000.0
+96.0,T1,425.0,1.0,2042.9999999999984,800000.0
+97.0,T1,425.0,1.0,2050.499999999998,700000.0
+98.0,T1,425.0,1.0,2056.999999999998,600000.0
+99.0,T1,425.0,1.0,2062.499999999998,500000.0
+100.0,T1,425.0,1.0,2066.999999999998,400000.0
+101.0,T1,425.0,1.0,2070.499999999998,300000.0
+102.0,T1,425.0,1.0,2072.999999999998,200000.0
+103.0,T1,425.0,1.0,2074.499999999998,100000.0
+104.0,T1,425.0,1.0,2074.999999999998,0.0
