+++ project/validationTests/3S1T/dc/exports/substations.csv
@@ -1,0 +1,4 @@
+id,feed_nodes,return_node,emf,internal_resistance,rectifier_type
+SS0,N1,N0,900.0,0.1,DIODE
+SS1,N2,N0,900.0,0.1,DIODE
+SS2,N3,N0,900.0,0.1,DIODE
