+++ project/validationTests/3S1T/dc/results/electrical_application.csv
@@ -1,0 +1,11 @@
+time_s,project,scenario,base_hash,object_type,object_id,signal,value,unit,stage,iter,note
+00:00:00,0,900.000000,900.000000,900.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000
+00:00:01,1,900.000000,900.000000,900.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000
+00:00:02,2,887.905057,899.062587,899.993792,0.000000,0.000000,0.000000,0.000000,100000.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,100000.000000,0.000000,0.000000,0.000000,-100000.000000,-100000.000000,0.000000,0.000000
+00:00:03,3,873.329380,897.892332,899.986042,0.000000,0.000000,0.000000,0.000000,200000.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,200000.000000,0.000000,0.000000,0.000000,-200000.000000,-200000.000000,0.000000,0.000000
+00:00:04,4,869.166775,897.516340,899.983552,0.000000,0.000000,0.000000,0.000000,300000.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,300000.000000,0.000000,0.000000,0.000000,-300000.000000,-300000.000000,0.000000,0.000000
+00:00:05,5,859.005997,896.572259,899.977300,0.000000,0.000000,0.000000,0.000000,400000.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,400000.000000,0.000000,0.000000,0.000000,-400000.000000,-400000.000000,0.000000,0.000000
+00:00:06,6,848.903702,895.570080,899.970663,0.000000,0.000000,0.000000,0.000000,500000.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,500000.000000,0.000000,0.000000,0.000000,-500000.000000,-500000.000000,0.000000,0.000000
+00:00:07,7,838.947611,894.422658,899.963064,0.000000,0.000000,0.000000,0.000000,600000.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,600000.000000,0.000000,0.000000,0.000000,-600000.000000,-600000.000000,0.000000,0.000000
+00:00:08,8,829.079244,893.188090,899.954888,0.000000,0.000000,0.000000,0.000000,700000.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,700000.000000,0.000000,0.000000,0.000000,-700000.000000,-700000.000000,0.000000,0.000000
+00:00:09,9,819.415564,891.750182,899.945365,0.000000,0.000000,0.000000,0.000000,800000.000000,0,0,0,0.000000,0.000000,0.000000,0.000000,0.000000,0.000000,800000.000000,0.000000,0.000000,0.000000,-800000.000000,-800000.000000,0.000000,0.000000
